Managing streaks

+++ Streaker.API/Controllers/StreaksController.cs
@@ -42,16 +42,7 @@ namespace Streaker.API.Controllers
             var userId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier).Value;
             var userCanAccess = await _streaksService.CheckUserAuthorityAsync(userId, id);
             if (!userCanAccess)
-                return new ObjectResult(new ApiResponse()
-                {
-                    Errors = new()
-                    {
-                        [""] = ["You cannot access this streak."]
-                    }
-                })
-                {
-                    StatusCode = StatusCodes.Status403Forbidden,
-                };
+                return Forbid();
 
             return Ok(new ApiResponse<StreakDetailsDto>(streak));
         }
@@ -82,7 +73,12 @@ namespace Streaker.API.Controllers
                     }
                 });
 
-            await _streaksService.UpdateStreakAsync(streakUpdateDto);
+            var userId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier).Value;
+            var userCanAccess = await _streaksService.CheckUserAuthorityAsync(userId, id);
+            if (!userCanAccess)
+                return Forbid();
+
+            await _streaksService.UpdateStreakAsync(id, streakUpdateDto);
             return Created();
         }
 
@@ -100,11 +96,13 @@ namespace Streaker.API.Controllers
                     }
                 });
 
+            var userId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier).Value;
+            var userCanAccess = await _streaksService.CheckUserAuthorityAsync(userId, id);
+            if (!userCanAccess)
+                return Forbid();
+
             await _streaksService.DeleteStreakAsync(id);
-            return new ObjectResult(new ApiResponse())
-            {
-                StatusCode = StatusCodes.Status204NoContent
-            };
+            return NoContent();
         }
     }
 }
